refactor(Ladder): Ladder Builder 객체 제거 -> Ladder 메서드로 추가
